+++ macos/Classes/AppLinksMacosPlugin.swift
@@ -1,6 +1,12 @@
 import Cocoa
 import FlutterMacOS
 
+public class AppLinks {
+  static public let shared = AppLinksMacosPlugin()
+
+  private init() {}
+}
+
 public class AppLinksMacosPlugin: NSObject, FlutterPlugin, FlutterStreamHandler, FlutterAppLifecycleDelegate {
   private var eventSink: FlutterEventSink?
   private var initialLink: String?
@@ -8,17 +14,28 @@ public class AppLinksMacosPlugin: NSObject, FlutterPlugin, FlutterStreamHandler,
   private var initialLinkSent = false
 
   public static func register(with registrar: FlutterPluginRegistrar) {
-    let instance = AppLinksMacosPlugin()
+    let instance = AppLinks.shared
 
     let methodChannel = FlutterMethodChannel(name: "com.llfbandit.app_links/messages", binaryMessenger: registrar.messenger)
     registrar.addMethodCallDelegate(instance, channel: methodChannel)
 
     let eventChannel = FlutterEventChannel(name: "com.llfbandit.app_links/events", binaryMessenger: registrar.messenger)
     eventChannel.setStreamHandler(instance)
-    
+
     registrar.addApplicationDelegate(instance)
   }
   
+  public func getUniversalLink(_ userActivity: NSUserActivity) -> URL? {
+    guard userActivity.activityType == NSUserActivityTypeBrowsingWeb,
+        let url = userActivity.webpageURL,
+        let _ = NSURLComponents(url: url, resolvingAgainstBaseURL: true) else {
+        return nil
+    }
+    
+    return url
+  }
+
+  // Custom URL schemes
   public func handleWillFinishLaunching(_ notification: Notification) {
     NSAppleEventManager.shared().setEventHandler(
       self,
@@ -27,7 +44,20 @@ public class AppLinksMacosPlugin: NSObject, FlutterPlugin, FlutterStreamHandler,
       andEventID: AEEventID(kAEGetURL)
     )
   }
-
+  
+  // Universal links
+//  public override func application(_ application: NSApplication,
+//                                   continue userActivity: NSUserActivity,
+//                                   restorationHandler: @escaping ([any NSUserActivityRestoring]) -> Void) -> Bool {
+//
+//    guard let url = getUniversalLink(userActivity) else {
+//      return false
+//    }
+//    
+//    handleLink(link: url.absoluteString)
+//    
+//    return false
+//  }
 
   public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
     switch call.method {
@@ -36,7 +66,7 @@ public class AppLinksMacosPlugin: NSObject, FlutterPlugin, FlutterStreamHandler,
         break
       case "getLatestAppLink":
         result(latestLink)
-        break      
+        break
       default:
         result(FlutterMethodNotImplemented)
         break
@@ -72,7 +102,7 @@ public class AppLinksMacosPlugin: NSObject, FlutterPlugin, FlutterStreamHandler,
     }
   }
 
-  private func handleLink(link: String) {
+  public func handleLink(link: String) {
     latestLink = link
 
     if (initialLink == nil) {